Added support for spruce transition to shrub, tweaks to gram->shrub, rework of basalarea calcs

+++ include/GraminoidTundra.h
@@ -39,8 +39,8 @@ private:
 	static const double*	_pStartAgeParms;
 	static EStartAgeType	_startAgeType;
 	static double*			_pIntegral;
-	static std::vector<float>			_rollingTempMean;
-	static std::vector<float>			_rollingSWIMean;
+	static std::vector<double>			_rollingTempMean;
+	static std::vector<double>			_rollingSWIMean;
 
 //Functions
 public:
